Refactor audio merging to be in a hook instead of in the AudioVisualizer component.
This refactor allows for the AudioVisualizer component to be way more simpler it receives
a single audio stream and visualizes it at all times. All the merging logic is now in a hook.

This also allows for the audio stream to be recorded in a way that is more reliable since
we only need to record the merged stream.